+++ crates/runner-nordic/src/board.rs
@@ -29,15 +29,13 @@ pub mod fpc2534;
 pub mod gpio;
 pub mod led;
 pub mod platform;
-#[cfg(feature = "radio-ble")]
-pub mod radio;
 mod rng;
 pub mod timer;
 #[cfg(feature = "uart")]
 pub mod uart;
 pub mod usb;
 #[cfg(feature = "_vendor")]
-mod vendor;
+pub mod vendor;
 
 impl board::Api for Board {
     fn try_event() -> Option<board::Event<Board>> {
@@ -65,8 +63,6 @@ impl board::Api for Board {
     type Gpio = gpio::Impl;
     type Led = led::Impl;
     type Platform = platform::Impl;
-    #[cfg(feature = "radio-ble")]
-    type Radio = radio::Impl;
     type Rng = rng::Impl;
     type Storage = crate::storage::Storage;
     type Timer = timer::Impl;